requirements.txt
tensorflow==2.14.0
Flask==3.0.0
Pillow==10.1.0
numpy==1.24.3
gunicorn==21.2.0
Werkzeug==3.0.0
